+++ packages/gamut/src/ProgressBar/__tests__/ProgressBar-test.tsx
@@ -43,4 +43,16 @@ describe('ProgressBar', () => {
 
     expect(wrapped.text()).toEqual('50%');
   });
+
+  it('uses an svg when given a pattern', () => {
+    const wrapped = renderComponent({ pattern: 'diagonalStripesRegular' });
+
+    expect(wrapped.find('svg').length).toBe(1);
+  });
+
+  it('does not use an svg when not given a pattern', () => {
+    const wrapped = renderComponent();
+
+    expect(wrapped.find('svg').length).toBe(0);
+  });
 });