+++ packages/gamut/src/index.tsx
@@ -35,6 +35,7 @@ export * from './NotificationList';
 export * from './NotificationList/NotificationIcon';
 export * from './NotificationList/NotificationItem';
 export * from './Overlay';
+export * from './Pattern';
 export * from './ProgressBar';
 export * from './RadialProgress';
 export * from './SkipToContent';