+++ packages/gamut-illustrations/CHANGELOG.md
@@ -3,6 +3,23 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+### [0.4.1](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-illustrations@0.4.0...@codecademy/gamut-illustrations@0.4.1) (2021-01-13)
+
+**Note:** Version bump only for package @codecademy/gamut-illustrations
+
+
+
+
+
+## [0.4.0](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-illustrations@0.3.12...@codecademy/gamut-illustrations@0.4.0) (2021-01-12)
+
+
+### Features
+
+* **Icons:** Add video player movie icon ([def3a59](https://github.com/Codecademy/client-modules/commit/def3a59d9c253330bc4703f2c46ded80a1079bac))
+
+
+
 ### [0.3.12](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-illustrations@0.3.11...@codecademy/gamut-illustrations@0.3.12) (2021-01-08)
 
 **Note:** Version bump only for package @codecademy/gamut-illustrations